+++ VEnt_Omega/Vent_box/Vent_box/main.h
@@ -64,8 +64,8 @@ int millis;
 int interval;
 int delta_U;
 int delta_F;
-int delta_T;
-int delta_H;
+float delta_T;
+float delta_H;
 int timer_M;
 int timer_U;
 unsigned char arm;       //пуск ручной
